Create Templates container
