+++ src/subapps/projects/components/Templates/TemplatesList.less
@@ -1,0 +1,8 @@
+.templates-list {
+  &__cards {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+    margin: 2em 0;
+  }
+}
