+++ src/subapps/projects/components/Templates/TemplatesList.tsx
@@ -1,0 +1,20 @@
+import * as React from 'react';
+
+import TemplateCard from './TemplateCard';
+
+import './TemplatesList.less';
+
+const TemplatesList: React.FC<{ templates: any[] }> = ({ templates }) => {
+  return (
+    <div className="templates-list">
+      <h2>Select Template</h2>
+      <div className="templates-list__cards">
+        {templates.length
+          ? templates.map(template => <TemplateCard template={template} />)
+          : 'No templtes found for this project.'}
+      </div>
+    </div>
+  );
+};
+
+export default TemplatesList;
